+++ packages/cli-core/src/format/format.test.ts
@@ -1,26 +1,26 @@
 import * as os from 'node:os';
 import * as assert from 'node:assert';
 import { after, before, describe, it } from 'node:test';
-import { format } from './format.js';
+import { Format, format } from './format.js';
 import { $, blue, bold, cyan, green, underline } from 'kleur/colors';
 
 void describe('format', () => {
-  void it('should format amplify command with yarn', () => {
-    const command = 'help';
-    const binaryRunner = 'yarn';
-    const expectedOutput = cyan(`yarn amplify help`);
-    const actualOutput = format.runner(binaryRunner).amplifyCommand(command);
-    assert.strictEqual(actualOutput, expectedOutput);
+  void it('should format amplify command with yarn', { concurrency: 1 }, () => {
+    const formatter = new Format('yarn');
+    assert.strictEqual(
+      formatter.normalizeBackendCommand('help'),
+      cyan('yarn amplify help')
+    );
   });
 
   void it('should return error for empty amplify command', () => {
-    const binaryRunner = 'yarn';
     assert.throws(
       () => {
-        format.runner(binaryRunner).amplifyCommand('');
+        format.normalizeBackendCommand('');
       },
-      Error,
-      'Command cannot be empty'
+      {
+        message: 'The command must be non-empty',
+      }
     );
   });
 
@@ -101,13 +101,13 @@ void describe('format when terminal colors disabled', async () => {
 
   void it('prints plain command', () => {
     const message = 'hello';
-    const coloredMessage = format.runner('yarn').amplifyCommand(message);
+    const coloredMessage = format.normalizeBackendCommand(message);
 
     assert.strictEqual(
       coloredMessage.includes('\x1b['),
       false,
       'Color codes should not be present'
     );
-    assert.strictEqual(coloredMessage, 'yarn amplify hello');
+    assert.strictEqual(coloredMessage, 'npx amplify hello');
   });
 });