+++ packages/cli-core/src/package-manager-controller/pnpm_package_manager_controller.ts
@@ -23,7 +23,6 @@ export class PnpmPackageManagerController extends PackageManagerControllerBase {
     super(
       cwd,
       'pnpm',
-      'pnpm',
       ['init'],
       'install',
       fsp,