+++ packages/integration-tests/src/test-e2e/deployment.test.ts
@@ -88,7 +88,7 @@ void describe('deployment tests', { concurrency: testConcurrencyLevel }, () => {
             await amplifyCli(
               [
                 'generate',
-                'config',
+                'outputs',
                 '--branch',
                 testBranch.branchName,
                 '--app-id',