+++ packages/model-generator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @aws-amplify/model-generator
 
+## 0.7.0
+
+### Minor Changes
+
+- d0f1452: bump codegen and amplify data cdk construct, schema generator dependencies
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [c2c8910]
+  - @aws-amplify/deployed-backend-client@0.4.1
+
 ## 0.6.0
 
 ### Minor Changes